requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
concurrent-log-handler==0.9.25
Deprecated==1.2.14
Flask==3.0.3
Flask-Limiter==3.8.0
Flask-Login==0.6.3
flask-restx==1.3.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
portalocker==2.10.1
Pygments==2.18.0
pymodbus==2.5.3
pyserial==3.5
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.19.1
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
pydantic==2.11.3
python-dotenv==1.1.0
Flask-SQLAlchemy==3.1.1
cachetools==5.5.2
pytest==8.3.5
pytest-html==4.1.1
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
pyjwt==2.10.1
PyYAML==6.0.2
psutil==6.1.1